requirements.txt

Flask==2.0.3
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.6.0
inflection==0.5.1

itsdangerous==2.1.0
Jinja2==3.0.3

keras==2.8.0
Keras-Preprocessing==1.1.2


mypy-extensions==0.4.3
numpy==1.21.5
Pillow==9.0.1

requests-oauthlib==1.3.1

tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109

webencodings==0.5.1
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
